requirements.txt
amqp==2.5.2
asgiref==3.2.7
bilibiliupload==0.2.2
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-filter==2.2.0
django-timezone-field==4.0
djangorestframework==3.11.0
flower==0.9.4
humanize==0.5.1
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.8
Markdown==3.2.1
pyasn1==0.4.8
PyExecJS==1.5.1
python-crontab==2.4.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.8
vine==1.3.0
zipp==3.1.0
https://github.com/FortuneDayssss/BilibiliUploader/releases/download/0.0.1/bilibiliuploader-0.0.1.tar.gz
